Create_QESD_monthly_fire_scar_record.py
Using a template document, update the string 'data_string' so that it has the relevant month and year inserted.
Script will be modified to read the relevant month and year as variables and change MMMM and YYYY from these variables. The update data string is then made into an API end point to create a new record in QESD catalogue.